Also remember to normalize/strip whitespaces in text query [#18]
