+++ lib/shimmer.rb
@@ -8,6 +8,7 @@ require "shimmer/keyboard_driver"
 require "shimmer/mouse_driver"
 require "shimmer/javascript_bridge"
 require "shimmer/javascript_evaluation_error"
+require "shimmer/javascript_expressions"
 
 module Shimmer
 end
